Refactored NoteListing
