+++ app/javascript/src/App.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Route, BrowserRouter as Router, Switch } from "react-router-dom";
 
 import Sidebar from "components/Common/Sidebar";
-import NoteListing from "components/Notes/NoteListing";
+import NoteListing from "components/Notes";
 import "lib/day";
 
 const App = () => (
@@ -13,10 +13,6 @@ const App = () => (
       <div className="relative flex h-screen flex-grow flex-col overflow-auto">
         <Switch>
           <Route component={NoteListing} path="/" />
-          <Route path="/form-elements" />
-          <Route path="/formik" />
-          <Route path="/overlays" />
-          <Route path="/layouts" />
         </Switch>
       </div>
     </div>
